remove old alpha instance metadata service
