+++ packages/backend-defaults/report-alpha.api.md
@@ -5,7 +5,6 @@
 ```ts
 import { ActionsRegistryService } from '@backstage/backend-plugin-api/alpha';
 import { ActionsService } from '@backstage/backend-plugin-api/alpha';
-import { InstanceMetadataService } from '@backstage/backend-plugin-api/alpha';
 import { ServiceFactory } from '@backstage/backend-plugin-api';
 
 // @public (undocumented)
@@ -18,13 +17,6 @@ export const actionsRegistryServiceFactory: ServiceFactory<
 // @public (undocumented)
 export const actionsServiceFactory: ServiceFactory<
   ActionsService,
-  'plugin',
-  'singleton'
->;
-
-// @alpha @deprecated (undocumented)
-export const instanceMetadataServiceFactory: ServiceFactory<
-  InstanceMetadataService,
   'plugin',
   'singleton'
 >;
