+++ packages/backend-defaults/src/alpha/index.ts
@@ -13,6 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 export { actionsRegistryServiceFactory } from './entrypoints/actionsRegistry';
 export { actionsServiceFactory } from './entrypoints/actions';
-export { instanceMetadataServiceFactory } from './entrypoints/instanceMetadata';
